Use "import/export" syntax for GGRC.Date utils
