+++ src/ggrc-client/js/components/datepicker/datepicker.js
@@ -7,6 +7,7 @@ import {
   inViewport,
   isInnerClick,
 } from '../../plugins/ggrc_utils';
+import {DATE_FORMAT} from '../../plugins/utils/date-util';
 
 (function (can, GGRC, moment) {
   'use strict';
@@ -73,7 +74,7 @@ import {
         this.picker.datepicker('setDate', null);
         this.attr('date', null);
       },
-      MOMENT_DISPLAY_FMT: GGRC.Date.MOMENT_DISPLAY_FMT,
+      MOMENT_DISPLAY_FMT: DATE_FORMAT.MOMENT_DISPLAY_FMT,
     }),
 
     events: {
@@ -84,9 +85,9 @@ import {
         let maxDate;
         let date;
         let options = {
-          dateFormat: GGRC.Date.PICKER_ISO_DATE,
+          dateFormat: DATE_FORMAT.PICKER_ISO_DATE,
           altField: this.element.find('.datepicker__input'),
-          altFormat: GGRC.Date.PICKER_DISPLAY_FMT,
+          altFormat: DATE_FORMAT.PICKER_DISPLAY_FMT,
           onSelect: this.viewModel.onSelect.bind(this.viewModel),
         };
 
@@ -127,7 +128,7 @@ import {
           // into account to not end up with a different date. Ideally this
           // should never happen, but that would require refactoring the way
           // Date objects are created throughout the app.
-          return moment(date).format(GGRC.Date.MOMENT_ISO_DATE);
+          return moment(date).format(DATE_FORMAT.MOMENT_ISO_DATE);
         } else if (this.isValidDate(date)) {
           return date;
         }
@@ -136,7 +137,7 @@ import {
       },
 
       isValidDate: function (date) {
-        return moment(date, GGRC.Date.MOMENT_ISO_DATE, true).isValid();
+        return moment(date, DATE_FORMAT.MOMENT_ISO_DATE, true).isValid();
       },
 
       /**
@@ -172,7 +173,7 @@ import {
           // into account to not end up with a different date. Ideally this
           // should never happen, but that would require refactoring the way
           // Date objects are created throughout the app.
-          date = moment(date).format(GGRC.Date.MOMENT_ISO_DATE);
+          date = moment(date).format(DATE_FORMAT.MOMENT_ISO_DATE);
         }
         date = moment.utc(date);
 
@@ -198,9 +199,9 @@ import {
 
         if (val) {
           val = val.trim();
-          valF = moment.utc(val, GGRC.Date.MOMENT_DISPLAY_FMT, true);
+          valF = moment.utc(val, DATE_FORMAT.MOMENT_DISPLAY_FMT, true);
           valISO = valF.isValid() ?
-            valF.format(GGRC.Date.MOMENT_ISO_DATE) :
+            valF.format(DATE_FORMAT.MOMENT_ISO_DATE) :
             null;
         }
         return valISO;
@@ -215,7 +216,7 @@ import {
           if (currentDateObj < updated) {
             this.viewModel.attr(
               '_date',
-              moment.utc(updated).format(GGRC.Date.MOMENT_DISPLAY_FMT));
+              moment.utc(updated).format(DATE_FORMAT.MOMENT_DISPLAY_FMT));
           }
         }
       },
